[contacts_api] move addressbook to it's own file
